requirements.txt
# Core dependencies for Streamlit Cloud deployment
streamlit==1.29.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
keras==3.12.0
jax[cpu]==0.4.20
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
Pillow==10.1.0
pyttsx3==2.90